fix: dataloader

+++ src/app/data/datasets.py
@@ -205,7 +205,7 @@ class BYUCustomDataset(Dataset):  # type: ignore[misc]
             )  # could return vxs there. see get_locs_n_vxs definition
             mask = torch.zeros_like(tomogram)
             if isinstance(zyx, torch.Tensor):
-                mask[zyx[:, 0], zyx[:, 1], zyx[:, 2]] = 1.0
+                mask[zyx[:, 0].to(torch.int), zyx[:, 1].to(torch.int), zyx[:, 2].to(torch.int)] = 1.0
             else:
                 zyx = torch.tensor(zyx)
             return {
